pyproject.toml
[project]
name = "eli-domain-monitor"
version = "0.1.0"
description = "CLI tool that monitors domains for expiration dates, status changes, and nameserver information"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.3",
    "dnspython>=2.3.0",
    "pyyaml>=6.0",
    "rich>=13.3.5",
    "pytest>=8.4.2",
    "whoisit @ git+https://github.com/meeb/whoisit.git@main",
]

[project.optional-dependencies]
test = [
    "pytest>=8.4.2",
]





[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (may require network access)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
